requirements.txt
numpy
numpy-stl==3.0.1
pyserial